Clean up analyze_source to the point where scripts can be written individually using its functionality without needing to hack in code directly in analyze_source.py
